+++ src/Pim/Bundle/EnrichBundle/Form/Type/AttributeOptionValueType.php
@@ -15,6 +15,17 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
  */
 class AttributeOptionValueType extends AbstractType
 {
+    /** @var string */
+    protected $dataClass;
+
+    /**
+     * @param string $dataClass
+     */
+    public function __construct($dataClass)
+    {
+        $this->dataClass = $dataClass;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -51,7 +62,7 @@ class AttributeOptionValueType extends AbstractType
      */
     protected function addFieldValue(FormBuilderInterface $builder)
     {
-        $builder->add('value', null, array('required' => false));
+        $builder->add('value', null, ['required' => false]);
     }
 
     /**
@@ -60,9 +71,9 @@ class AttributeOptionValueType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(
-            array(
-                'data_class' => 'Pim\Bundle\CatalogBundle\Entity\AttributeOptionValue'
-            )
+            [
+                'data_class' => $this->dataClass,
+            ]
         );
     }