+++ src/Pim/Bundle/EnrichBundle/Form/Type/GroupTypeType.php
@@ -16,6 +16,17 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
  */
 class GroupTypeType extends AbstractType
 {
+    /** @var string */
+    protected $dataClass;
+
+    /**
+     * @param string $dataClass
+     */
+    public function __construct($dataClass)
+    {
+        $this->dataClass = $dataClass;
+    }
+
     /**
      * {@inheritdoc}
      */
@@ -25,12 +36,12 @@ class GroupTypeType extends AbstractType
         $builder->add(
             'label',
             'pim_translatable_field',
-            array(
+            [
                 'field'             => 'label',
                 'translation_class' => 'Pim\\Bundle\\CatalogBundle\\Entity\\GroupTypeTranslation',
                 'entity_class'      => 'Pim\\Bundle\\CatalogBundle\\Entity\\GroupType',
                 'property_path'     => 'translations'
-            )
+            ]
         );
         $builder->addEventSubscriber(new DisableFieldSubscriber('code'));
     }
@@ -41,9 +52,9 @@ class GroupTypeType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(
-            array(
-                'data_class' => 'Pim\Bundle\CatalogBundle\Entity\GroupType'
-            )
+            [
+                'data_class' => $this->dataClass,
+            ]
         );
     }