+++ src/Pim/Bundle/EnrichBundle/Form/Type/ImageType.php
@@ -15,20 +15,31 @@ use Symfony\Component\OptionsResolver\OptionsResolverInterface;
  */
 class ImageType extends AbstractType
 {
+    /** @var string */
+    protected $dataClass;
+
+    /**
+     * @param string $dataClass
+     */
+    public function __construct($dataClass)
+    {
+        $this->dataClass = $dataClass;
+    }
+
     /**
      * {@inheritdoc}
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('file', 'file', array('required' => false))
+            ->add('file', 'file', ['required' => false])
             ->add(
                 'removed',
                 'checkbox',
-                array(
+                [
                     'required' => false,
                     'label'    => 'Remove media',
-                )
+                ]
             )
             ->add('id', 'hidden')
             ->add('copyFrom', 'hidden');
@@ -40,9 +51,9 @@ class ImageType extends AbstractType
     public function setDefaultOptions(OptionsResolverInterface $resolver)
     {
         $resolver->setDefaults(
-            array(
-                'data_class' => 'Pim\Bundle\CatalogBundle\Model\ProductMedia'
-            )
+            [
+                'data_class' => $this->dataClass,
+            ]
         );
     }